update to new monocypher api

+++ src/hashing/blake2b.rs
@@ -3,7 +3,6 @@
 //! [Official documentation](https://monocypher.org/manual/hash)
 
 use ffi;
-use libc::size_t;
 use std::mem;
 
 /// Simple function to hash the input data.
@@ -18,7 +17,6 @@ use std::mem;
 pub fn easy(data: &[u8]) -> [u8; 64] {
     general(data, b"")
 }
-
 
 /// Function to hash the input data with an additional key.
 ///
@@ -36,9 +34,9 @@ pub fn general(data: &[u8], key: &[u8]) -> [u8; 64] {
             hash.as_mut_ptr() as *mut u8,
             64,
             key.as_ptr(),
-            key.len() as size_t,
+            key.len() as u64,
             data.as_ptr(),
-            data.len() as size_t,
+            data.len() as u64,
         );
         hash.assume_init()
     }
@@ -64,8 +62,12 @@ impl Context {
     pub fn new(key: &[u8]) -> Context {
         unsafe {
             let mut ctx = mem::MaybeUninit::<ffi::crypto_blake2b_ctx>::uninit();
-            ffi::crypto_blake2b_general_init(ctx.as_mut_ptr() as *mut ffi::crypto_blake2b_ctx,
-                                             64, key.as_ptr(), key.len());
+            ffi::crypto_blake2b_general_init(
+                ctx.as_mut_ptr() as *mut ffi::crypto_blake2b_ctx,
+                64,
+                key.as_ptr(),
+                key.len() as u64,
+            );
             Context(ctx.assume_init())
         }
     }
@@ -74,7 +76,7 @@ impl Context {
     #[inline]
     pub fn update(&mut self, data: &[u8]) {
         unsafe {
-            ffi::crypto_blake2b_update(&mut self.0, data.as_ptr(), data.len());
+            ffi::crypto_blake2b_update(&mut self.0, data.as_ptr(), data.len() as u64);
         }
     }
 
@@ -82,7 +84,7 @@ impl Context {
     #[inline]
     pub fn finalize(&mut self) -> [u8; 64] {
         unsafe {
-            let mut hash= mem::MaybeUninit::<[u8; 64]>::uninit();
+            let mut hash = mem::MaybeUninit::<[u8; 64]>::uninit();
             ffi::crypto_blake2b_final(&mut self.0, hash.as_mut_ptr() as *mut u8);
             hash.assume_init()
         }
@@ -91,8 +93,8 @@ impl Context {
 
 #[cfg(test)]
 mod test {
-    use hex;
     use super::*;
+    use hex;
 
     #[test]
     fn blake2b_incremental() {
